Remove cache clear in BeforeAll/AfterAll methods
It is possible for multiple BeforeAll and AfterAll lifecycle methods in
JUnit to get called on the same thread as a running test, causing cached
drivers to get quit while still in use.

This change adds a documented warning to that API and prevents our test
of JUnit 5 parallelism from falling into that trap.

For context, see related issues and PRs:
* https://github.com/apache/groovy-geb/pull/181
* https://github.com/apache/groovy-geb/issues/188
* https://github.com/apache/groovy-geb/issues/288
* https://github.com/apache/groovy-geb/pull/299
* https://github.com/apache/groovy-geb/pull/301